requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
boto3==1.38.46
botocore==1.38.46
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==3.0.0
Django==5.2.3
django-appconf==1.1.0
django-debug-toolbar==5.2.0
django-imagekit==5.0.0
django-storages==1.14.6
gevent==25.5.1
goodconf==6.1.2
greenlet==3.2.3
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
jmespath==1.0.1
oauthlib==3.3.1
packaging==25.0
pilkit==3.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
zope.event==5.1
zope.interface==7.2
